Refactored hierarchy and activity delegation.
