+++ app/src/main/AndroidManifest.xml
@@ -22,5 +22,8 @@
 
         </activity>
 
+        <activity android:name=".experience.ExperienceActivity"
+                  android:screenOrientation="portrait"/>
+
     </application>
 </manifest>
