+++ app/src/main/java/com/comic_con/museum/ar/experience/launchar/LaunchArFragment.kt
@@ -1,4 +1,4 @@
-package com.comic_con.museum.ar.content
+package com.comic_con.museum.ar.experience.launchar
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -7,12 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import com.comic_con.museum.ar.R
 
-class ContentFragment: Fragment() {
+class LaunchArFragment: Fragment() {
 
     private var rootView: View? = null
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val rootView = inflater.inflate(R.layout.fragment_content, container, false)
+        val rootView = inflater.inflate(R.layout.fragment_launch_ar, container, false)
         this.rootView = rootView
 
         return rootView
