+++ app/src/main/res/menu/bottom_nav.xml
@@ -3,12 +3,12 @@
 
     <item
             android:id="@+id/progress"
-            android:icon="@drawable/baseline_camera_black_48"
+            android:icon="@drawable/baseline_star_black_48"
             android:title="@string/nav_progress"/>
 
     <item
             android:id="@+id/content"
-            android:icon="@drawable/baseline_camera_black_48"
+            android:icon="@drawable/baseline_view_list_black_48"
             android:title="@string/nav_content"/>
 
     <item
